improve web

+++ db/migrate/20150331073638_alter_column_donation_amounte.rb
@@ -1,5 +1,5 @@
 class AlterColumnDonationAmounte < ActiveRecord::Migration
   def change
-  	change_column :donations, :amount, :float
+  	change_column :donations, :amount, :integer
   end
 end
